requirements.txt
anyio==4.3.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.2.2
exceptiongroup==1.2.1
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
python-dotenv==1.0.1
python-telegram-bot==13.13
pytz==2024.1
six==1.16.0
sniffio==1.3.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
